+++ internal/provider/resource_iosxe_interface_port_channel_test.go
@@ -43,6 +43,9 @@ func TestAccIosxeInterfacePortChannel(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "vrf_forwarding", "VRF1"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ipv4_address", "192.0.2.1"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ipv4_address_mask", "255.255.255.0"))
+	if os.Getenv("C9000V") != "" {
+		checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "switchport", "false"))
+	}
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ip_access_group_in", "1"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ip_access_group_in_enable", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ip_access_group_out", "1"))
@@ -68,6 +71,18 @@ func TestAccIosxeInterfacePortChannel(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ipv6_addresses.0.prefix", "2224:DB8::/32"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ipv6_addresses.0.eui_64", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "arp_timeout", "2147"))
+	if os.Getenv("C9000V") != "" {
+		checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ip_arp_inspection_trust", "true"))
+	}
+	if os.Getenv("C9000V") != "" {
+		checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ip_arp_inspection_limit_rate", "1000"))
+	}
+	if os.Getenv("C9000V") != "" {
+		checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "spanning_tree_link_type", "point-to-point"))
+	}
+	if os.Getenv("C9000V") != "" {
+		checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ip_dhcp_snooping_trust", "true"))
+	}
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -125,6 +140,9 @@ func testAccIosxeInterfacePortChannelConfig_all() string {
 	config += `	vrf_forwarding = "VRF1"` + "\n"
 	config += `	ipv4_address = "192.0.2.1"` + "\n"
 	config += `	ipv4_address_mask = "255.255.255.0"` + "\n"
+	if os.Getenv("C9000V") != "" {
+		config += `	switchport = false` + "\n"
+	}
 	config += `	ip_access_group_in = "1"` + "\n"
 	config += `	ip_access_group_in_enable = true` + "\n"
 	config += `	ip_access_group_out = "1"` + "\n"
@@ -156,6 +174,18 @@ func testAccIosxeInterfacePortChannelConfig_all() string {
 	config += `		eui_64 = true` + "\n"
 	config += `	}]` + "\n"
 	config += `	arp_timeout = 2147` + "\n"
+	if os.Getenv("C9000V") != "" {
+		config += `	ip_arp_inspection_trust = true` + "\n"
+	}
+	if os.Getenv("C9000V") != "" {
+		config += `	ip_arp_inspection_limit_rate = 1000` + "\n"
+	}
+	if os.Getenv("C9000V") != "" {
+		config += `	spanning_tree_link_type = "point-to-point"` + "\n"
+	}
+	if os.Getenv("C9000V") != "" {
+		config += `	ip_dhcp_snooping_trust = true` + "\n"
+	}
 	config += `	depends_on = [iosxe_restconf.PreReq0, ]` + "\n"
 	config += `}` + "\n"
 	return config