+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `iosxe_errdisable` resource and data source
 - Add `iosxe_line` resource and data source
+- Add `spanning_tree_link_type` and `ip_dhcp_snooping_trust` attributes to `iosxe_interface_port_channel` resource and data source
 
 ## 0.4.0